+++ LibCarla/source/carla/client/detail/Simulator.h
@@ -123,6 +123,20 @@ namespace detail {
 
     /// @}
     // =========================================================================
+    /// @name Required files related methods
+    // =========================================================================
+    /// @{
+
+    bool SetFilesBaseFolder(const std::string &path);
+
+    std::vector<std::string> GetRequiredFiles(const std::string &folder = "", const bool download = true) const;
+
+    void RequestFile(const std::string &name) const;
+
+    std::vector<uint8_t> GetCacheFile(const std::string &name, const bool request_otherwise) const;
+
+    /// @}
+    // =========================================================================
     /// @name Garbage collection policy
     // =========================================================================
     /// @{
@@ -427,6 +441,10 @@ namespace detail {
       _client.SetActorAutopilot(vehicle.GetId(), enabled);
     }
 
+    void ShowVehicleDebugTelemetry(Vehicle &vehicle, bool enabled = true) {
+      _client.ShowVehicleDebugTelemetry(vehicle.GetId(), enabled);
+    }
+
     void SetLightsToVehicle(Vehicle &vehicle, const rpc::VehicleControl &control) {
       _client.ApplyControlToVehicle(vehicle.GetId(), control);
     }
@@ -651,6 +669,8 @@ namespace detail {
     const GarbageCollectionPolicy _gc_policy;
 
     SharedPtr<Map> _cached_map;
+
+    std::string _open_drive_file;
   };
 
 } // namespace detail