+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * Changed the resolution of the cached map in Traffic Manager from 0.1 to 5 meters
   * Fixed import sumo_integration module from other scripts
+  * CARLA now is built with Visual Studio 2019 in Windows
   * Fixed bug causing the RoadOptions at the BehaviorAgent to not work as intended
   * Upgrading to Unreal Engine 4.26
   * Added Lincoln 2020 vehicle dimensions for CarSim integration
@@ -14,12 +15,14 @@
   * Added 'check_lidar_bb' util script
   * Added optional flag to `client.replay_file()` `replay_sensors` to enable or disable the replaying the sensors
   * Improved manual_control: now cameras are set in relation with car size
+  * Client required files are sent from the server to a local cache (OpenDRIVE, Traffic Manager...)
   * Added CHRONO library for vehicle dynamics simulation (https://projectchrono.org/)
     - Supported JSON vehicle definition
     - Unsupported collision dynamics
   * Added performance benchmarking section to documentation
   * Added API functions to traffic light actor `get_effect_waypoints()`, `get_light_boxes()` and `get_opendrive_id()`
   * Added API functions to world `get_traffic_lights_from_waypoint()` and `get_traffic_lights_in_junction`
+  * Added API function to enable car telemetry
   * CARLA is compatible with the last RoadRunner nomenclature for road assets
   * Fixed a bug when importing a FBX map with some **_** in the FBX name
   * Extended make import process for applying road painter materials (carla art tool)